lxd/network/openvswitch: Rename to ovs
Sponsored-by: Luizalabs (https://luizalabs.com)
License: Apache-2.0
