+++ lxd/network/ovs/schema/ovs/autoattach.go
